+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.6.0
+
+- Added mock axios constructor
+
 ## 1.5.0
 
 - Added Methods